+++ apps/simulator/src/components/blocks/index.ts
@@ -1,4 +1,4 @@
-export * from "./RegularQuestions/SATA/RegularSATAQuestionaire";
+export * from "./RegularQuestions/SATA/RegSATAQuestionaireBlock";
 export * from "./RegularQuestions/MCQ/MCQBlocks";
 export * from "./CaseStudy/CaseStudyContainer";
 export * from "./CaseStudy/CaseStudyQuestions/MRSNQuestionaire";
