+++ packages/utils/hooks/index.ts
@@ -3,5 +3,4 @@ export * from "./useApi";
 export * from "./usePreviousValue";
 export * from "./useCustomAction";
 export * from "./useSessionStorage";
-export * from "./useCustomErrorhandling";
 export * from "./useFormFocusOnError";
